Add the distinction between actual form errors and errors displayed
A form, if any of its fields are invalid it should be considered invalid even if the errors are not displayed for example due to the isNewModel property that does not show all field errors at the beginning. So I created the distinction between the actual errors of the form, and the errors that are being displayed in the form.