Enable lodash 'paths' feature
According to the documentation, most features were turned off
by default to reduce the package size. In order to use your desired
feature, you needed to opt in to it.

'paths' feature was set to false. According to documentation at
, `paths` feature offers "Deep property path support for methods like _.get, _.has, & _.set".
If the feature is not enabled, deep property fetching will fail since
the methods are replaced with a _.noop method that always returns `undefined`.

In my usecase, I have fields with models like 'foo.bar' which means
that these fields will not work since we are able to set the value,
but not get the value.

Therefore, I have enabled this option so that we can be sure that
vfg.abstractField's value property will always be working whether the
model provided is a deep or shallow path.