Use Promise success and failure handlers for .validate
Previously, the validate() function returned a promise which
would resolve() if there were no validation errors, and reject(errors)
if there were errors. However, when calling this validate method, we called it as the callback
of the "fields-validation-trigger" event.

When this promise is rejected because of validation errors, we
do not have a .catch method to deal with the rejection. This causes
it to get "stuck" and dies in the console. This is unwanted behavior.

The javascript engine will track such unhandled rejections and generate
global errors with the event "unhandledrejection".

This fills the console with "errors" which aren't actual execution errors.

The solution I had here was to update the code to always use resolve. If there
are errors, we resolve with the errors, if there are no errors, we resolve with
undefined/null.

The solution I implemented here is to include empty functions success and failure
handlers.